backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
mssql-django
pyodbc
